moved SerialPortExt trait to seperate file
